circuit: small reorg
