+++ circuit/src/lib.rs
@@ -3,6 +3,7 @@ extern crate alloc;
 
 pub mod builder;
 pub mod circuit;
+pub mod errors;
 pub mod expr;
 pub mod op;
 pub mod tables;
@@ -12,8 +13,9 @@ pub mod utils;
 
 // Re-export public API
 pub use builder::{CircuitBuilder, CircuitBuilderError};
-pub use circuit::Circuit;
+pub use circuit::{Circuit, CircuitField};
+pub use errors::CircuitError;
 pub use expr::{Expr, ExpressionGraph};
 pub use op::{FakeMerklePrivateData, NonPrimitiveOp, NonPrimitiveOpPrivateData, Prim};
-pub use tables::{CircuitError, CircuitRunner, FakeMerkleTrace, Traces};
+pub use tables::{CircuitRunner, FakeMerkleTrace, Traces};
 pub use types::{ExprId, NonPrimitiveOpId, WitnessAllocator, WitnessId};
